roomCodeMessage in component
